Add junit-vintage-engine as a test dependency.
Update to Spring Boot as best I can tell starts to use JUnit 5, but tests in project are JUnit 4.

Added the junit-vintage-engine to the list of project dependencies in the pom.xml file. This addition will support running of JUnit Vintage tests including JUnit 3 and JUnit 4 tests in the project.